[API] 페이지네이션, 팔로워 등 API 개선 기능들

+++ src/main/java/threads/server/domain/like/repository/LikeCommentRepository.java
@@ -8,5 +8,5 @@ import java.util.Optional;
 public interface LikeCommentRepository extends JpaRepository<LikeComment, Long>, LikeRepository<LikeComment> {
     Optional<LikeComment> findByUserIdAndCommentId(Long userId, Long commentId);
 
-    Integer countByCommentId(Long id);
+    Long countByCommentId(Long id);
 }
